TeachAI Claude v.2/requirements.txt
# Основные библиотеки
openai>=1.3.0
ipywidgets>=7.7.0
jupyter>=1.0.0
python-dotenv>=0.19.0

# Библиотеки для обработки данных
pandas>=1.3.0
numpy>=1.20.0

# Библиотеки для визуализации (используются в примерах)
matplotlib>=3.4.0
seaborn>=0.11.0
